Stop doing ping because it's not really needed
I was launching vbar inside .xinitrc, which meant that it wouldn't have
an X display to work with, which in turn meant that vbar would freeze up
and not process any update commands.

I added the ping to get around this, but it turns out that's not
a working solution.

Instead, launch vbar inside your window manager startup code so that
you're sure there's an X display available.